fixing cross-layer dependency between dto and AbstractInjectionModification
