+++ src/main/java/org/gridsuite/modification/modifications/AbstractInjectionModification.java
@@ -7,25 +7,22 @@
 package org.gridsuite.modification.modifications;
 
 import com.powsybl.commons.report.ReportNode;
-import com.powsybl.commons.report.TypedValue;
 import com.powsybl.iidm.network.Injection;
 import com.powsybl.iidm.network.extensions.Measurement;
 import com.powsybl.iidm.network.extensions.Measurements;
 import com.powsybl.iidm.network.extensions.MeasurementsAdder;
 import org.gridsuite.modification.dto.InjectionModificationInfos;
+import org.gridsuite.modification.utils.MeasurementUtils;
 import org.gridsuite.modification.utils.ModificationUtils;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.UUID;
 
 /**
  * @author Ayoub LABIDI <ayoub.labidi at rte-france.com>
  */
 public abstract class AbstractInjectionModification extends AbstractModification {
 
-    private static final String VALUE = "value";
-    private static final String VALIDITY = "validity";
     protected final InjectionModificationInfos modificationInfos;
 
     protected AbstractInjectionModification(InjectionModificationInfos modificationInfos) {
@@ -49,8 +46,8 @@ public abstract class AbstractInjectionModification extends AbstractModification
         }
         // measurements update
         List<ReportNode> reports = new ArrayList<>();
-        upsertMeasurement(measurements, Measurement.Type.ACTIVE_POWER, pValue, pValidity, reports);
-        upsertMeasurement(measurements, Measurement.Type.REACTIVE_POWER, qValue, qValidity, reports);
+        MeasurementUtils.upsertMeasurement(measurements, Measurement.Type.ACTIVE_POWER, pValue, pValidity, reports);
+        MeasurementUtils.upsertMeasurement(measurements, Measurement.Type.REACTIVE_POWER, qValue, qValidity, reports);
         // report changes
         ReportNode estimSubReportNode = null;
         if (!reports.isEmpty()) {
@@ -60,47 +57,4 @@ public abstract class AbstractInjectionModification extends AbstractModification
         return estimSubReportNode;
     }
 
-    public static void upsertMeasurement(Measurements<?> measurements, Measurement.Type type, Double value, Boolean requestedValidity, List<ReportNode> reports) {
-        if (value == null && requestedValidity == null) {
-            return;
-        }
-        String measurementType = (type == Measurement.Type.ACTIVE_POWER ? "Active power" : "Reactive power") + " measurement ";
-        Measurement measurement = getExistingMeasurement(measurements, type);
-        if (measurement != null) { // update measurement
-            if (value != null) {
-                double oldValue = measurement.getValue();
-                measurement.setValue(value);
-                if (reports != null) {
-                    reports.add(ModificationUtils.buildModificationReport(oldValue, value, measurementType + VALUE, TypedValue.INFO_SEVERITY));
-                }
-            }
-            if (requestedValidity != null) {
-                boolean oldValidity = measurement.isValid();
-
-                ModificationUtils.updateMeasurementValidity(measurement, requestedValidity);
-                if (reports != null) {
-                    reports.add(ModificationUtils.buildModificationReport(oldValidity, requestedValidity, measurementType + VALIDITY, TypedValue.INFO_SEVERITY));
-                }
-            }
-        } else {
-            var measurementAdder = measurements.newMeasurement().setId(UUID.randomUUID().toString()).setType(type);
-            if (value != null) {
-                measurementAdder.setValue(value);
-                if (reports != null) {
-                    reports.add(ModificationUtils.buildModificationReport(null, value, measurementType + VALUE, TypedValue.INFO_SEVERITY));
-                }
-            }
-            if (requestedValidity != null) {
-                measurementAdder.setValid(requestedValidity);
-                if (reports != null) {
-                    reports.add(ModificationUtils.buildModificationReport(null, requestedValidity, measurementType + VALIDITY, TypedValue.INFO_SEVERITY));
-                }
-            }
-            measurementAdder.add();
-        }
-    }
-
-    public static Measurement getExistingMeasurement(Measurements<?> measurements, Measurement.Type type) {
-        return measurements.getMeasurements(type).stream().findFirst().orElse(null);
-    }
 }
